% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{getUnisensCustomAttributes}
\alias{getUnisensCustomAttributes}
\title{Get Unisens Custom Attributes}
\usage{
getUnisensCustomAttributes(unisensFolder)
}
\arguments{
\item{unisensFolder}{Unisens Folder}
}
\value{
hash map of all custom attributes
}
\description{
Get Unisens Custom Attributes
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
getUnisensCustomAttributes(unisensPath)
}
