% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{getUnisensMeasurementId}
\alias{getUnisensMeasurementId}
\title{Get Unisens Measurement Id Time}
\usage{
getUnisensMeasurementId(unisensFolder)
}
\arguments{
\item{unisensFolder}{Unisens Folder}
}
\value{
string measurement id
}
\description{
Get Unisens Measurement Id Time
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
getUnisensMeasurementId(unisensPath)
}
