% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{getUnisensSignalSampleCount}
\alias{getUnisensSignalSampleCount}
\title{Get Unisens Signal Sample Count}
\usage{
getUnisensSignalSampleCount(unisensFolder, id)
}
\arguments{
\item{unisensFolder}{Unisens Folder}

\item{id}{ID of the signal entry}
}
\value{
Long
}
\description{
Get Unisens Signal Sample Count
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
getUnisensSignalSampleCount(unisensPath, 'ecg.bin')
}
