% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensSignalEntry}
\alias{readUnisensSignalEntry}
\title{Read Unisens Signal Entry}
\usage{
readUnisensSignalEntry(
  unisensFolder,
  id,
  binReadChunkSize = 2^16,
  startIndex = 1,
  endIndex = getUnisensSignalSampleCount(unisensFolder, id)
)
}
\arguments{
\item{unisensFolder}{String containing path to Unisens folder.}

\item{id}{String containing ID of the signal entry.}

\item{binReadChunkSize}{Interger of chunk size of read values for bin file format, default: 2^16.}

\item{startIndex}{Integer of the value-index on which the read process starts, default: 1.}

\item{endIndex}{Integer of the value-index on which the read process ends, default: last Index of file.}
}
\value{
DataFrame.
}
\description{
Read Unisens Signal Entry
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
readUnisensSignalEntry(unisensPath, 'ecg.bin')
}
