% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensStartTime}
\alias{readUnisensStartTime}
\title{Read Unisens Start Time}
\usage{
readUnisensStartTime(unisensFolder)
}
\arguments{
\item{unisensFolder}{Unisens Folder}
}
\value{
POSIXct unisens start time
}
\description{
Read Unisens Start Time
}
\examples{
readUnisensStartTime('./unisensR/extdata/unisensExample')
}
