% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensValuesEntry}
\alias{readUnisensValuesEntry}
\title{Read Unisens Values Entry}
\usage{
readUnisensValuesEntry(unisensFolder, id)
}
\arguments{
\item{unisensFolder}{Unisens Folder}

\item{id}{ID of the values entry.}
}
\value{
DataFrame.
}
\description{
Read Unisens Values Entry
}
\examples{
readUnisensValuesEntry('./unisensR/extdata/unisensExample', 'rr.csv')
}
