% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensSignalEntry}
\alias{readUnisensSignalEntry}
\title{Read Unisens Signal Entry}
\usage{
readUnisensSignalEntry(unisensFolder, id, binReadChunkSize = 2^16)
}
\arguments{
\item{unisensFolder}{Unisens Folder}

\item{id}{ID of the signal entry}

\item{binReadChunkSize}{chunk size of read values for bin file format, default: 2^16}
}
\value{
DataFrame.
}
\description{
Read Unisens Signal Entry
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
readUnisensSignalEntry(unisensPath, 'ecg.bin')
}
