% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensSignalEntry}
\alias{readUnisensSignalEntry}
\title{Read Unisens Signal Entry}
\usage{
readUnisensSignalEntry(unisensFolder, id)
}
\arguments{
\item{unisensFolder}{Unisens Folder}

\item{id}{ID of the signal entry.}
}
\value{
DataFrame.
}
\description{
Read Unisens Signal Entry
}
\examples{
readUnisensSignalEntry('../UnisensR/inst/extdata/unisensExample', 'ecg.bin')
}
