% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisensR.R
\name{readUnisensSignalEntry}
\alias{readUnisensSignalEntry}
\title{Read Unisens Signal Entry}
\usage{
readUnisensSignalEntry(
  unisensFolder,
  id,
  endIndex = getUnisensSignalSampleCount(unisensFolder, id),
  binReadInChunks = FALSE,
  binReadChunkSize = 2^16,
  startIndex = 1
)
}
\arguments{
\item{unisensFolder}{String containing path to Unisens folder.}

\item{id}{String containing ID of the signal entry.}

\item{endIndex}{Integer of the value-index on which the read process ends, default: last Index of file.}

\item{binReadInChunks}{Boolean determines if the reading process of binary files is done in chunks.
This could be useful if you run into memory limits when reading big files. default: FALSE.}

\item{binReadChunkSize}{Interger of chunk size of read values for bin file format, default: 2^16.}

\item{startIndex}{Integer of the value-index on which the read process starts, default: 1.}
}
\value{
DataFrame.
}
\description{
Read Unisens Signal Entry
}
\examples{
unisensPath <- system.file('extdata/unisensExample', package = 'unisensR', mustWork = TRUE)
readUnisensSignalEntry(unisensPath, 'ecg.bin')
}
